requirements.txt
simpy>=4.0.1
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
asyncio-mqtt>=0.13.0
paho-mqtt>=1.6.0
scipy>=1.10.0
scikit-learn>=1.3.0
jupyter>=1.0.0
pytest>=7.0.0
pyyaml>=6.0
colorlog>=6.7.0
pulp>=3.0.0